+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     * `UUID`: `0.3.0` -> `0.3.1`
 * `Common`:
     * New property `MPPFile#withoutSlashAtTheEnd`
+    * Extension `clamp` has been deprecated
+    * New extension `Iterable#diff`
 
 ## 0.5.24